+++ tests/testthat/test-mlp-multinomial.R
@@ -17,7 +17,7 @@ df$y <- factor(df$y)
 test_that("multinomial mlp", {
   skip_if_not(torch::torch_is_installed())
   skip_if(packageVersion("rlang") < "1.0.0")
-  skip_on_os("linux")
+  skip_on_os(c("windows", "linux", "solaris"))
 
   expect_snapshot({
     set.seed(1)
@@ -39,8 +39,7 @@ test_that("multinomial mlp", {
 test_that("class weights - mlp", {
   skip_if_not(torch::torch_is_installed())
   skip_if(packageVersion("rlang") < "1.0.0")
-  skip_on_os("linux")
-
+  skip_on_os(c("windows", "linux", "solaris"))
 
   expect_snapshot({
     set.seed(1)